Fix internal dependencies for packaging + remove floating-duration (#19)
cargo package wasn't working on the physx crate as we only specified path-dependencies.

Fixed it so we specify both path and version dependencies which is the standard way to handle this. It will use the path dependencies when building the repo and the version dependency when packaging and for the published crate.