Replace BodyHandle with ActorHandle and ArticulationHandle
This is a partial solution to a flaw in the current crate API regarding actor deletion.
Previously, a BodyHandle was created for both articulations and static/dynamic actors.
However, the remove_body function found in the physics module called remove_articulation,
regardless of the type actually pointed to by the BodyHandle. This lead to actors never being removed and a PhysX warning being printed.

I've addressed this by replacing BodyHandle with ActorHandle and ArticulationHandle, which are handled separately.
ActorHandle also has a type field that allows handles to be removed from internally tracked lists.
I'm not very confident about some of the changes around articulations, and have not used articulations in PhysX,
so I could use some thorough review/testing in that area. I think we could really benefit from some tests. :)
I by no means think this is a perfect or final solution, but want to start the feedback process early.

This is a breaking change and warrants a bump in the crate's major verison, if using semantic versioning.