Build for x86_64-linux-android with generated/unix/ code
After first running structgen (cross-compiled with the latest Android NDK) on
an x86_64-linux-android Android emulator I confirmed that the generated
code is identical to the existing `generated/unix` code.

Being able to build for x86_64-linux-android is particularly useful for
being able to run applications under an Android emulator.